produce abstraction for project level generator
